+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## branch `pre-filled-amounts`
+* Add ability to pre-fill amounts form
+
 ## 2.2.2 (2019-02-28)
 * Fix plugin on WordPress 4